requirements.txt
# --- Core AI & LLM Generation ---
# Required for PyTorch backend
torch>=2.0.0
# Hugging Face Transformers for loading the Phi-2 model
transformers>=4.37.0
# Required for device_map="auto" and efficient model loading
accelerate>=0.26.0
# Required for 4-bit quantization (BaseAgent configuration)
bitsandbytes>=0.41.0
# Often required by transformers for mathematical operations
scipy

# --- Audio Processing & Transcription ---
# OpenAI's Whisper model for speech-to-text
openai-whisper
# For downloading audio from YouTube
yt-dlp
# Note: You also need FFmpeg installed on your system (not a python package)

# --- Presentation Generation ---
# For creating .pptx files
python-pptx
# For downloading images to insert into slides
requests

# --- External APIs ---
# Wrapper for Pexels Stock Photo API
pexels-api

# --- Utilities ---
# Optional but recommended for managing API keys in .env files
python-dotenv